+++ index.html
@@ -19,6 +19,7 @@
             <div class="legend-keys"><span class="No-change status"></span> No Change</div>
         </div>
     </div>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
+    
 </body>
 </html>
