+++ lang-and-F.html
@@ -40,6 +40,7 @@
             </div>
         </section> -->
 
-    <script src="../index.js"></script>
+    <script src="languageAndFrameworks.js"></script>
+    <script type="module" src="pages.js"></script>
 </body>
 </html>
