+++ languageAndFrameworks.js
@@ -1,0 +1,7 @@
+
+fetch("https://tech-radar-api.herokuapp.com/tech-radar")
+  .then((response) => response.json())
+  .then((data) => {
+    console.log(data)
+    
+  });
