+++ platforms.html
@@ -12,37 +12,8 @@
         <h1>Platforms</h1>
     </header>
 
-        <!-- <aside>
-            <div id="Adopt" class="Adopt">
-                <h1>Adopt</h1>
-            </div>
-            <div id="Trial" class="Trial">
-                <h1>Trial</h1>
-            </div>
-            <div id="Assess" class="Assess">
-                <h1>Assess</h1>
-            </div>
-            <div id="Hold" class="Hold">
-                <h1>Hold</h1>
-            </div>
-
-        </aside>
-
-        <section >
-            <div class="quadrant-section platforms" id="grid">
-
-            </div> 
-            
-            <div class="evalPhases">
-                <p>Adopt</p>
-                <p>Trial</p>
-                <p>Assess</p>
-                <p>Hold</p>
-            </div>
-        </section>
        
-
-     -->
      <script src="../index.js"></script>
+     <script type="module" src="pages.js"></script>
 </body>
 </html>
