+++ techniques.html
@@ -43,5 +43,6 @@
 
     -->
     <script src="../index.js"></script> 
+    <script type="module" src="pages.js"></script>
 </body>
 </html>
